Keep reference to Device objects so gc does not destroy them
Subscribe and unsubscribe events of taurus.core.util.event.EventGenerator
is based on callback and data objects uniqueness and match.
ValueBuffer and ValueReferenceBuffer objects use taurus.Device class objects
as data. If we don't keep references to them, these can be garbage collected.
Keep them in a list between subscribe and unsubscribe of the MeasurementGroup.