Add mechanism to select shelve backend for MS environment
ndbm backend does not work correctly in the way Sardana uses shelve.
This backend does not store data on disk when process receives SIGKILL or
SIGABRT. gnu and dumb however works much better.

Add MS_ENV_SHELVE_BACKEND to sardanacustomseetings in order to select between
gnu and dumb backends. In case it is not specified, first try to use gnu
and if not available use dumb.