Do not detect events on MeasurementGroup's LatencyTime and NbStarts
Due to a bug in PyTango (or Tango) we need to be very carefull with not
properly configured event system. Subscription to an attribute in stateless
mode which constantly fails may collide with destruction of Attribute/DeviceProxy.
More details in tango-controls/pytango#302. To be on the safe side don't
check event criteria for LatencyTime neither NbStarts. While in the case of the
first one we would like to always receive the events, even if the change
is extreamly low, the second one has sense only to send events if the difference
is 1. But for the moment we don't even check the NbStarts criteria.

Fixes #1182.