Fix WrongParam message showed in spock
WrongParam message, for example due to object not found, are showed in spock
as unknown sardana exception instead of showing a verbose message.
This is due to the fact that spock shows just the high level exception message.
Since Py3 we raise exception from lower level exceptions (see 77a7aa91) then
we must also pass the exception message to the high level exception. Do it.