Remove other documentation versions
In the last 2 years, the other_versions documentation was never really used.
Furthermore it is polluted with many documentation version for branches which
were uploaded to upstream but are not really considered as Sardana versions.
This may confuse users and in order to fix it manual removal is necessary.

The idea is that in the future, other documentation versions will be
hosted as dedicated GitHub repos and exported to GitHub Pages.

Then other_verions.rst will be reintroduced again, but this time as a static page.