Fix #1237: Macro functions with result don't return any result
Macro functions are just instances of the same MacroFunction class.
For every macro function call, the macroserver creates an instance
of MacroFunction.
The constructor of MacroFunction is already overwriting the values
of param_def and result_def so promoting hasResult() from
classmethod to method, should be sufficient to allow each macro
function to have distinct results.