Do not clear operation in acquisition action_loop
OperationConext wrongly cleared twice - within the action_loop and when
the OperationContext exits. In the case of fast acquisitions this may cause
that when leaving the previous operation context we clear the operation
context of the next one. Call set_state_info as the finish_hook so it will
be called when the operation context is cleared already.