Do not clear operation in motion action_loop
OperationConext is wrongly cleared twice - within the action_loop and when
the OperationContext exits. In the case of fast motions this may cause
that when leaving the previous operation context we clear the operation
context of the next one. Call set_state_info as the finish_hook so it will
be called when the operation context is cleared already.

Also protect the set_state_info with the lock.