Add back. comp. for value_ref_* parameters
Measurement groups created with expconf with Sardana prior to Jul20 release
has wrongly configured value_ref_enabled/value_ref_pattern parameters
for non-referable channels. This is now protected when setting new configuration,
and it raises unwanted exception when memorized attribute is recovered.
For backwards compatibility recognize when server is starting and in this case
just warn the user and raise the exception when setting this parameter by user.