Wait longer for DS to start in tests
It has been observed that when switching to execution of tests by pytest
sometimes the DS started by tests are taking more then 10 s to start.
Usually the start time is 3-4 s. This was especially reproducible when
executing all tests, then the workspace/sardana/src/sardana/tango/pool/test/test_Motor.py was failing. When failing we observed that it takes 11 s.