Fix problem of measurement group acquisitions
Two consecutive measurement group acquisitions executed very fast can
overlap. The software trigger of the next acquisition may be skipped cause
the previous one is still in progress.

Re-introduce the threading.Event for protecting the acquisition action
(it was already used in the past). On the first software trigger wait for
the previous acquisition to finish.