Do not use state to determine Value quality
Value quality was determined based on the channel's state. Since Value
events are pushed before the State events the state had to be updated
pre-maturely to correctly determine the Value quality. This pre-mature
state update was not propagating event and was just updating the channel's
internal state. This however was making the measurement group state to
switch to ON too early. The first channel's state update with propagation
was finding all the channels with internal state set to ON making the
measurement group state change to ON as well. This is causing bugs in fast
repeated acquisitions.

Hack it and use the *acquiring* flag which is independent of the state.
Set to to True on start and to False when the acquisition has finished.
Also reset it to False when leaving the operation just in case there was
some unhandled error or the user stopped the acquisition.