Avoid extra state readout at the end of acquisition
Acquisition's action_loop exists the state poll loop when none of the
elements is in Moving or Running state. This indicates that the acquisition
has finished. This last state readout should be propagated to the
channels. Avoid the extra state readout which actually could change the
end acquisition state. This also introduces a small performance improvement.