Ensure order of moveables is preserved in Motion object
Move of motion object sends motors to wrong positions. This happens
when using moveables from different Pools and is not 100 % reproducible.

This is due to the fact that we rely on simple dictionaries
(their order is not preserved in Python 3.5) when constructing
the Motion object from the list of moveables requested by the user.
Further move requests on that object assume that the positions
correspond to the order of moveables requested at the Motion
object creation time. Currently the order may be lost.

Use OrderedDict instead to avoid such problems.