Open msenvironment shelve file always with 'c' flag
Empty environment files used with dumb backend fails to
open with 'w' flag. This does not happen for gdbm backend.

Avoid this problem by always using 'c' flag, which according to
docs [1]: "Open database for reading and writing, creating it if it doesn’t exist".

Tested and it does not harm when the environment shelve files is not empty - its
content is preserved.

Fixes #1425.

[1] https://docs.python.org/3/library/dbm.html#dbm.open