add sardana_init_hook
Add sardana_init_hook to SardanaDevice for allowing the subscribe to pool's elements attribute in the init of the MacroServer.

Based on the response of Michal to https://gitlab.com/tango-controls/pytango/-/merge_requests/419. Modify the run_tango_server to replicate the behaviour of post_init_callback found in PyTango High-level API which is invoked before server enters event loop. Fixes #674.