(fix) Avoid already finished macro being aborted at MS exit
Macro which was already finished is considered as _running
macro_ and at the MacroServer (MS) exit the running macro gets
aborted. This abort is unnecessary and causes the MS to crash
on Windows.

- Introduce the concept of _last macro_.
- When the macro finishes, update the _last macro_ information
and at the same time clear the _running macro_ - this avoids
the macro aborting.
- Maintain the backwards compatibility when reading the macro
data from the door: if it is read during the macro execution
it reports the data of the macro currently being run and as
soon as the macro finished it returns the data of the macro
which just finished execution.