(test) increase timeout for qtspock tests
On Windows it was observed that the qtspock tests that
have timeout set to 5 s sporadically fail.
The ones with the timeout set to 10 s works well.
This happened only when executed within the testsuite,
when executed separately it was not observed.
As an immediate solution, unify all timeouts in all tests
to 10 s.