(fix) sardana GUIs launchers on Windows
During the Jan21 release it was observed that the Sardana GUIs launchers,
when sardana is installed on Windows with pip, does not launch
the applications at all.
In the previous releases Sardana was never tested on Windows when installing with
pip, but with MSI, and in this case the launchers were working correctly.
Since version Jan21, the recommendation is to use Sardana on Windows with pip.
In order to overcome this problem change the GUIs launchers from gui_script to
console_scripts entry points. Similar solution was already applied for Taurus
in taurus-org/taurus#541.