(fix) make PMTV relative move combobox accept only positive numbers
In the PMTV widget, one can introduce negative values in the relative move write
widget. It is ambiguous since the relative movement is commanded with a move negative
or move positive buttons - there is no need to add a sign to the delta displacement.

Set bottom limit to 0 for the validator. sys.float_info.min was also evaluated but it
still allows to move the motor with 0 delta.