fix: restore motor's velocity in continuous scans
8815c91 introduced a hack to avoid recursive decrease of
velocity in continuous scans. Most probably this hack was needed
to avoid some synchronization problems between the software
continuous scan threads (motion and acquisition are done in
separate threads).
When using Taurus 4 this hack is not applied
cause the attribute's range returns -inf and inf floats
instead of "Not specified" string and we don't fall into the
except clause anymore.
Refactor the `CScan.get_max_top_velocity()` to consider only
Taurus 4 and above case (we don't support Taurus < 4 anymore)
and maintain the hack.