fix: do not reserve instruments in GScan
GScan reserve instruments for the scanning macro when it simply
looks for the instruments information e.g. full names. This makes
the MacroExecutor to stop the instruments when the macro is
being stopped as it would do with any other reserved object.

Do not reserve instruments cause it does not make sense to stop
them when stopping the macro.