add: pool_controllers property to PoolAction and use it internally
PoolAction maintains reference to pool controllers and its elements
involved in the action. For simple actions e.g. PoolMotion
this is enough to internally operate directly on the data structure
keeping these references. However complex actions e.g. PoolAcquisition
which is a composition of other sub-actions, it is more convenient to use
pool_controllers property which could be overridden by the sub-class.
Add the property and use it internally.