add: override pool_controllers property in PoolAcquisition
PoolAcquisition is a composition of other sub-actions. In order to be able to
interrogate the MeasurementGroup about its state, it is necessary to consider
as pool controllers and its elements all the controllers and elements of the
sub-actions. Override the pool_controllers property to obtain the global
data structure. Fixes #1316.