fix: allow to start PoolElement only when it is in On or Alarm states
PoolElement can can try to start when its state is different than
Moving. This tries to start elements in Fault state as well
what can hang the server, for example, in the case of
the measurement group. See discussion in #1316. Allow to start
the PoolElement only if it is in On or Alarm state.