add: release hung objects when releasing hung macro
Macro can be interrupted by stopping/aborting. When the abort
have not worked, then the macro can be released on the user
request. This however almost certainly will leave the element's
operation in the Pool run forever making it impossible to reuse
the element again.
Release the hung element when releasing the macro - this releases
the operation as well.