Add devcontainer for VS Code - Remote Containers
This adds a docker-compose with four containers, one for the sardana development with a clean conda environment specified in the file environment.yml and the other 3 related to have a Tango CS (tangodb, databaseds and tangotest).

By default the TANGO_HOST is set to databaseds:10000 but an external Tango DB could be used.