fix: completely clean client side after removing a measurement group
MeasurementGroup _proxy_ object resides on the client side after
removing it on the server. This leads to inconsistency when one
recreates the measurement group using the same name e.g. its list
of the channels corresponds to the list of the channels of the
previously deleted measurement group.

Explicitly remove the Taurus objects (device and its attributes)
from the factory so they do not interfere with the future attempts
to recreate objects with the same name.