fix: correctly update macro_pointer
#1559 introduced a bug which does not allow to abort macros executing
hooked macros. This is because a dedicated `_macro_pointer` is being
wrongly invalidated at the end of the `runMacro()` method.

Refactor the concept of _macro_pointer_ - use a property which points to the macro
at the `_macro_stack`. This avoids the problem of having to update
the `_macro_pointer`.

Deprecate `clearRunningMacro()` in favor of `clearMacroStack()`.