fix: update MeasurementGroup's Elements property when Configuration attr is written
MeasurementGroup's persistent information about elements is stored in two places:
1. Elements property
2. Configuration memorized attribute
The first one is set only once, at the MeasurementGroup creation time. The subsequent
changes of the configuration does not update the Elements property. Update it on the
successful write of the Configurtion attribute in order to maintain them in sync.