fix: hang of IPython when macro input gives timeout
When macro input gives timeout the Spock (IPython) gets hung completely.
It seems to be bug in IPython - see https://github.com/ipython/ipython/issues/12078.
Use prompt_toolkit.prompt() (a dependency of IPython) instead of input() and raise an
exception in the thread waiting for input when the timeout was reached. This apparently
works.

Also change the message in Spock when the timeout is reached.