fix: correctly extract taurus version
taurus for many years was allowing to retrieve its version from the
taurus.core.release module directly, even if this was not a recommended way.
taurus!1199 does not allow it anymore.

Change the way of obtaining the version number. In the future it will be
possible to use taurus.__version__ directly.