expconf: enable "apply" on changes during load
Expconf assumes that the local config is untouched while setting-up,
but this is no longer true now that we can, e.g., sanitize the pre-scan
snapshot config during load. Therefore, flag the local config as dirty
(i.e. allow to apply it) if some config changed in the loading process.