fix: Update name before other methods get called
The property name of the elem is used in the rename_element method to
assign the new name of the element, however,
it is not updated in the moment the method
is executed.

We changed the order of these events in order to
have the name updated in the following calls to
rename_element.