fix: avoid memory leak in continuous scan (ct type) macros
aNscan, which holds a hard reference to the CTScan, adds its
bound method as hook in the waypoint generator.
The enumeration object created from the waypoint generator
is then hard referenced in the CTScan. Hence, a cycle reference
is created between CTScan and aNscan objects
(we imagine that the last yielded waypoint must be kept in the
enumeration object).
Then due to unknown reasons, the scan memory can not be released
if this cycle reference exist (even if we believe that ta some
point the GC should collect it).

Follow the example of the timescan macro and TScan class
in order to not pass  _fill_missing_records() as hook.
Refactor the code and define the _fill_missing_record in the
CAcquisition class (base class of both CTScan and TScan).
This way the memory is correctly released for ascanct, a2scanct,
meshct, etc. macros.

More details in the #1664 (hotfix-3.1.2).