fix: meas group with undefined disabled channel
Disabled channels in the measurement group are
considered as not members of the group. They neither appear
in "element list" attribute nor the measurement group listens to their events.
Hence it is possible to undefine an experimental channel which is disabled.
This may lead to a situation when we try to enable again an undefined channel.
Consider this case by not accepting a configuration which contains
an undefined element. Also allow to acquire with a measurement group which
contains and undefined and disabled channel.