fix: restore of memorized conf when channel is undefined (2nd attempt)
Since 6ac7aa67 measurement groups does not accept configurations with undefined channels - an exception is raised. This collides with restoring configuration memorized values when a channel was undefined
while still present, but disabled, in measurement groups configurations.
Raise the said exception only if the channel is enabled in the
measurement group.
Previously d9ecbc8ca also tried to fix this issue but it introduced
a regression when disabling all channels in the measurement group.