fix: reduce memory leak of scans
Macros are executed in worker threads of a ThreadPool of size 10.
Each macro may end up being executed by a different thread.
In the investigation of #1095 we observed that memory allocated
in small chunks in different threads is not released to the OS.
Reduce the memory consumption by always executing macros
in the same thread (per door).