refactor: log information in mv instead of attaching to exception
Status of the motor group does not contain the original exception details raised during the state readout of the involved moveables.
Its status only contains a summary of moveables states.
Whenever we run a mv macro with multiple moveables i.e. we use a
motor group, and it fails during the start, then we loose these
details. Instead of early obtaining information (already on the motor
group), leave it to the macro resposability to obtain information
from each moveable separatelly. This way we get the missing
details. Also, this is more similar to how it is done in case of
problematic end of motion.