Clear exception context when formatting in controller
When an exception happens during the StateOne() call then
the PoolController formats its traceback in order to include it
in the status. When Sardana is executed as a standalone server,
and we read status inside of the except caluse, then the exception
raised during the StateOne() call will have the __context__
information filled. This information is not necessary and introduces
noise in the status. Clear it.
Note that traceback.format_exception() has a chain parameter which
when set to False also clears the __context__, but in this case it also
clears the __cause__ that we would like to maintain.